+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - Integrate https://github.com/pointfreeco/swift-snapshot-testing/pull/944/changes
 - Integrate https://github.com/pointfreeco/swift-snapshot-testing/pull/998/changes
 - Fix nested iOS WebView and SpriteKit snapshots
+- Fix iOS trait overrides on Xcode 26.6